+++ tests/Module/RootModuleFileConverterTest.php
@@ -116,7 +116,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $jsonData = (object) array(
             '$schema' => 'http://puli.io/schema/2.0/manager/module',
             'name' => 'my/application',
-            'path-mappings' => (object) array(
+            'resources' => (object) array(
                 '/app' => 'res',
             ),
             'bindings' => (object) array(
@@ -292,7 +292,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
     {
         $jsonData = (object) array(
             'name' => 'my/application',
-            'path-mappings' => (object) array(
+            'resources' => (object) array(
                 '/app' => 'res',
             ),
             'bindings' => (object) array(